+++ PivotView.Core/Rendering/PivotRendererItem.cs
@@ -17,10 +17,10 @@ public class PivotRendererItem
     /// <summary>
     /// Width / Height
     /// </summary>
-    public double AspectRatio =>
+    public float AspectRatio =>
         DataItem.ImageWidth == 0 || DataItem.ImageHeight == 0
             ? 1.0f
-            : (double)DataItem.ImageWidth / DataItem.ImageHeight;
+            : (float)DataItem.ImageWidth / DataItem.ImageHeight;
 
     public AnimatableProperty<RectangleF> Frame { get; } = new(default);
 }
